Continue work on transition from use of int32_t everywhere to use of size_t where that will be the right type to use in the C++ code.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3683 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/externs.h
@@ -619,7 +619,7 @@ extern const char *files_to_read[MAX_INPUT_FILES],
 extern int csl_argc;
 extern const char **csl_argv;
 extern bool fasl_output_file;
-extern int output_directory;
+extern size_t output_directory;
 
 #ifdef BUILTIN_IMAGE
 const unsigned char *binary_read_filep;
